remove print statement; converted int field to float field

+++ surf/models.py
@@ -4,5 +4,5 @@ from django.db import models
 class SurfReport(models.Model):
     captured_at = models.DateTimeField()
     local_time = models.DateTimeField()
-    min_swell = models.IntegerField()
-    max_swell = models.IntegerField()
+    min_swell = models.FloatField()
+    max_swell = models.FloatField()
